+++ nrf_802154/driver/src/mac_features/nrf_802154_csma_ca.h
@@ -63,8 +63,9 @@
  *
  * @retval true   The function always returns true for compatibility reasons
  */
-bool nrf_802154_csma_ca_start(const nrf_802154_frame_t                     * p_frame,
-                              const nrf_802154_transmit_csma_ca_metadata_t * p_metadata);
+nrf_802154_tx_error_t nrf_802154_csma_ca_start(
+    const nrf_802154_frame_t                     * p_frame,
+    const nrf_802154_transmit_csma_ca_metadata_t * p_metadata);
 
 /**
  *@}