+++ nrf_802154/driver/src/nrf_802154_request.h
@@ -205,9 +205,10 @@ bool nrf_802154_request_rssi_measurement_get(int8_t * p_rssi);
  * @retval  true   The transmission procedure was scheduled.
  * @retval  false  The driver could not schedule the transmission procedure.
  */
-bool nrf_802154_request_transmit_raw_at(const nrf_802154_frame_t                * p_frame,
-                                        uint64_t                                  tx_time,
-                                        const nrf_802154_transmit_at_metadata_t * p_metadata);
+nrf_802154_tx_error_t nrf_802154_request_transmit_raw_at(
+    const nrf_802154_frame_t                * p_frame,
+    uint64_t                                  tx_time,
+    const nrf_802154_transmit_at_metadata_t * p_metadata);
 
 /**
  * @brief Requests a call to @ref nrf_802154_delayed_trx_transmit_cancel.
@@ -269,8 +270,9 @@ bool nrf_802154_request_receive_at_scheduled_cancel(uint32_t id);
  * @param[in]  p_metadata  Pointer to metadata structure. Contains detailed properties of data
  *                         to transmit.
  */
-bool nrf_802154_request_csma_ca_start(const nrf_802154_frame_t                     * p_frame,
-                                      const nrf_802154_transmit_csma_ca_metadata_t * p_metadata);
+nrf_802154_tx_error_t nrf_802154_request_csma_ca_start(
+    const nrf_802154_frame_t                     * p_frame,
+    const nrf_802154_transmit_csma_ca_metadata_t * p_metadata);
 
 /**
  *@}